Restructure adapters functions
